requirements.txt
streamlit
pandas
python-dotenv
numpy
PyPDF2
langchain
langchain-google-cloud-sql-pg
langchain-google-vertexai
keybert
scikit-learn
sentence-transformers
matplotlib
ragas
google-cloud-aiplatform
vertexai
aiohttp
asyncio
pydantic
langchain-google-genai